Fixes #1056: create a new top-level extension module
