+++ app/extension/examples/log-example/src/main/resources/META-INF/syndesis/syndesis-extension-definition.json
@@ -1,0 +1,6 @@
+{
+  "tags": [
+    "feature",
+    "experimental"
+  ]
+}
