+++ app/extension/maven-plugin/src/main/java/io/syndesis/maven/RepackageIntegrationMojo.java
