+++ app/extension/maven-plugin/src/resources/META-INF/spring.factories
