+++ app/rest/pom.xml
@@ -52,7 +52,7 @@
     <junit.version>4.12</junit.version>
     <kubernetes.client.version>2.4.1</kubernetes.client.version>
     <maven-shared-utils.version>3.1.0</maven-shared-utils.version>
-    <maven-common-artifact-filters.version>1.4.0</maven-common-artifact-filters.version>
+    <maven-common-artifact-filters.version>1.4</maven-common-artifact-filters.version>
     <mockwebserver.version>0.0.15</mockwebserver.version>
     <okhttp-eventsource.version>1.3.2</okhttp-eventsource.version>
     <postgresql.version>9.1-901-1.jdbc4</postgresql.version>
@@ -66,8 +66,7 @@
     <spring-security.version>4.2.3.RELEASE</spring-security.version>
     <spring-social.version>1.1.4.RELEASE</spring-social.version>
     <undertow.version>1.4.21.Final</undertow.version>
-    <atlasmap.version>1.32.2</atlasmap.version>
-    <camel-atlasmap.version>1.32.2</camel-atlasmap.version>
+    <camel-atlasmap.version>1.32.1</camel-atlasmap.version>
     <apache-mime4j.version>0.6</apache-mime4j.version>
     <google-testing-compile.version>0.12</google-testing-compile.version>
     <google-testing-compile-guava.version>21.0</google-testing-compile-guava.version>
@@ -105,10 +104,8 @@
     <module>project-generator</module>
     <module>rest</module>
     <module>runtime</module>
-    <module>syndesis-annotation-processor</module>
     <module>syndesis-build-helper-maven-plugin</module>
     <module>syndesis-builder-image-generator</module>
-    <module>syndesis-maven-plugin</module>
     <module>verifier</module>
   </modules>
